FIX payment to work with updated payment module
Payment functionality has been updated to work with the 1.0.x codeline of
silverstripe/payment. This means that the return-url of any of the payment
gateways point back at the event register form, which handles the inbound
redirect from the payment source directly.

+++ code/forms/EventRegisterForm.php
@@ -76,7 +76,17 @@ class EventRegisterForm extends MultiForm {
 			return false;
 		}
 
-		parent::finish($data, $form);
+		$success = parent::finish($data, $form);
+		if ($success === false) {
+			return false;
+		}
+		
+		// see whether we've been redirected by the step (most typically with Payment steps)
+		$controller = Controller::curr();
+		/* @var $controller Controller */
+		if ($controller->redirectedTo()) {
+			return;
+		}
 
 		$this->session->delete();
